Change update widget logic
I realised it makes no sense to shift the other widgets up when
updating the zindex of an existing widget. This means that you would
never be able to move a widget on top of another one, because as you
updated its zindex, the zindexes of the ones above it would also
increase, to maintain the relative order.